requirements.txt
Django==1.6
python-instagram
